Remove Equal and IsOrdered from graph.EndpointPair
Originally, we were going to enforce that EndpointPair's `==`
cannot be used by making the type incomparable, but this caused
compilation problems because set.Set requires that elements are
comparable. We could have made set.Set accept any type, but that
would require making a new hash table implementation with user-
defined equals and hash code functions, which Go has no first-class
support for.

So now that EndpointPair.Equal is no longer needed, we are forced
to remove the reason why it exists: IsOrdered. Therefore, we have
to remove orderedness from EndpointPair entirely so that == runs
in a way that makes sense.