Partly migrate Makefile and Bash scripts to Go
This is an experiment to see how far I can go with using Go to
script everything, similar to Rust's approach to using an `x.py`
script.

This is done partly out of fun and partly to allow linting to run
tools concurrently without resorting to Bash trickery or
installing GNU parallel (though it may be adopted anyway if Nix is
adopted as a way of managing consistent versions of various tools).

Issue: #45